backend/requirements.txt
Django>=5.0,<6.0
djangorestframework
requests
django-cors-headers
psycopg2-binary
python-dotenv
gunicorn
whitenoise
drf-yasg
djangorestframework-simplejwt
django-filter
dj-database-url
python-decouple
